Add custom authentication function feature
This commit adds a new AuthFunc field to the AuthOptions struct.  The User and
Password fields are left in-place for backwards compatibility.  This construct
shouldn't break any user code unless they are using unlabeled struct literals,
in which case it should break on principle since you shouldn't do that on
structs outside of your control.